refactor(api) : restructure lib folder
- organize functions related to http api
